Fix module parsing if event name has a check mark
Let's talk about Dualis: This is a shitty piece of software. Now that we've
established a common basis, let's get to the bug.

Dualis does not distinguish between an event name and the modules result
row. Both rows have the same identifier. The bug was caused by a too
loose condition that was responsible for detecting the module's result
row. Let's pray that Dualis does not have more weird edge cases like
this.